+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Optimization:
 Changes:
 - Removed component limit in blueprint selection.
 - Improve blueprint import UX.
+- Redstone Link can now connect to Redstone Link in different Send/Receive mode.
 
 Bugfixes:
 - Fixed gauges may not save ingredient amount change if source gauge relocated.